Updated openstack/openstack Project: openstack/heat  a0b1916888dfdf585f41fa77f44567d46f673ad0
Move NoActionRequired exception back where it belongs

It makes no sense for an exception that is used for local flow control -
especially flow control between the Resource superclass and its
subclasses - to be defined in the heat.common.exceptions module where
user-facing exceptions whose definitions must be shared between the engine
and api are defined.

This patch moves it back to the original, correct location and adds
docstrings so people will actually know what it's for.

Technically both this and the original move were breaking changes to the
third-party Resource plugin API. This exception seems much less likely to
be in widespread use than UpdateReplace (c.f. preceding patch) though.

This reverts commit 5120f096d7ef8d14c6a0d763659d4c51d1f00345.

Change-Id: I73dc671cd7e6e37ecb0f13545c7fa93d58e6e8ff
Related-Bug: #1611104