Updated openstack/openstack Project: openstack/glance-specs  83ad95084a145634e10dee03835216c4c28ee1c0
api-docs: deprecate the Images v1 API

This patch updates the "narrative" Images v1 API docs for the
Newton release by adding a deprecation note to each page.  That
way the note will be seen even if someone uses a direct link to
one of the pages.

The Images v1 API was deprecated with commit
63e6dbb1eb006758fbcf7cae83e1d2eacf46b4ab.

Change-Id: I62315efc12ee05aea5f55d020006cbb2ad279876
Partial-bug: #1618495