Updated openstack/openstack Project: openstack/python-swiftclient  0ec6b7b162a7b98ca3fad515de17d8f8a88dda72
Make tempurl subcommand insist on whole number seconds

Previously the tempurl subcommand would dump a traceback
due to a TypeError if the seconds option was not an int
value. With this patch it will now return the same
error message as if the seconds option were negative or not
a number.

Also changes the error message to state that the seconds
option should be a "whole number" rather than a "positive
integer", since 0 is a valid value.

Change-Id: Ie940d470f2be8006aa8eb7fe242f092457aeae21
Closes-Bug: #1621817