Updated openstack/openstack Project: openstack/heat  a267fd006b80d11376225cf127624903076842e0
Refactor resource definition parsing in HOT/cfn

Parse each field in a resource definition individually instead of all in
one go. This allows the HOT and cfn formats to apply different parsing to
different fields - specifically in the future we want to parse the
condition field like a condition, using only the valid condition functions.

Change-Id: I5b864d241a5e16c09fcce30c40d634d9bb72e173