Updated openstack/openstack Project: openstack/nova  6228068b1bdbfa9994a7b13924c198eba6662be3
[placement] Add some tests ensuring unicode resource provider info

Add some gabbi tests which demonstrate that it is possible to create
and query a resource provider that has a name that is outside the
bounds of ascii. The tests using a 4byte wide utf-8 character are left
as xfails because not all installations of mysql will support it.

Also confirm that if a unicode characters (uri-encoded or not) in
the uuid part of a resource providers path will result in the
expected 404 and not explode.

Closes-Bug: #1621605
Change-Id: I3d906c3296aa28b595fcc23c448c1744972c319d