Updated openstack/openstack Project: openstack-infra/project-config  00b5802d02a0c83a8d1306fba6f76bf1d873bf16
update api-ref to be a ./ path, not a symbolic name

I think that because the cloned path is the same name as the upstream
module, pip is installing the upstream module not the cloned path. Fix
how we pip install os-api-ref to make this work.

Remove the -q until this is working, to help with debugging, we can
set -q again when we know it all works.

Change-Id: I4fb65bd7a47c61d9d91c0064a8d6852ca5326163