Updated openstack/openstack Project: openstack-infra/project-config  b2fe8c459345dd38e7c2c9abf2a2bc0ed811be19
add zuul-cloner location to $PATH

We still run some of the release tools by hand, and the dev systems
where we do that do not have zuul-cloner installed to the same location
as on the CI nodes. Add the CI location to the shell's path instead of
invoking zuul-cloner using the full path explicitly.

Change-Id: I9a87e3e7873ea00fc738cc2a21a26cca2ff49fab
Signed-off-by: Doug Hellmann <doug@doughellmann.com>