Updated openstack/openstack Project: openstack/neutron  1a5105183669b2c15b0b8be9e87bb01f11a6f686
Add release note for blueprint vlan-aware-vms

Now that the bulk of the code is in place, and we are about
to see the finish line, it is time to prepare the release
notes for this feature.

This patch refines docstrings that recently went in and
makes sure the code is in sync with the agent's devref.

Implements: blueprint vlan-aware-vms

Change-Id: Idb9379046aeb07354a170a361da777140a2a028e