Updated openstack/openstack Project: openstack/heat  5dc1084260c8a7a7f5ec97d0cdfc1fc08cd8f82e
Make get_attr consistent across template versions

Several changes to when get_attr returns None as a placeholder were made
to the original version of HOT but not reflected in later versions. The
missing changes were:

- I1b2b4f0e967695cf3297d93022cecec958e94b24 (The fix for bug 1333208.)

- I6dbaa7ee4e9d534c31823b4812efcb387c695a22 (The implementation of the
  blueprint convergence-lightweight-stack. The lack of this suggests
  that our stacks may be less lightweight in practice than we thought.)

- I2bfac4ce6a8806acc85c17466429d63a5d93d79f (A small change to the
  previous patch.)

- I84762de87032168f97cdd19533649851073b4e81 (Part of the fix for
  bug 1486281)

This change refactors the classes such that new and old versions of the
templates share a single implementation of this test, so there should be
no future opportunity for them to get out of sync.

Change-Id: I367ba6abe5c0176ff84ec676a42e9329cf84b48e
Closes-Bug: #1623707