Updated openstack/openstack Project: openstack-infra/project-config  16155962e6c6914b947842b0d15467f82428dc5c
tripleo: disable voting on gate-tripleo-ci-centos-7-nonha-multinode

We are in an outstanding situation where multinode job is broken.
Our team is focusing on bringing it back on stage but until then, we
need to revert some patches that require
gate-tripleo-ci-centos-7-nonha-multinode to be non-voting.

This patch will be reverted once situations comes back to normal.

Context:
https://bugs.launchpad.net/tripleo/+bug/1623891
https://bugs.launchpad.net/tripleo/+bug/1623606

Change-Id: I107b4bdd4edf4dc54d1dad70333630a19714070c