Updated openstack/openstack Project: openstack/openstack-manuals  1602444280306134182fc2d980a7d985606de08c
Add marker files in the root of documents

This is needed for publishing in the future:
This is so that we can create a list of directories to ignore when we
rsync documentation builds to a publishing site. Some builds are
published as subdirectories underneath the results of superior
builds.

This allows those superior builds to be rsynced without removing the
other builds that were published under them.

Change-Id: I41257f4436ac7d030725bef5e6390390002cbe16