Updated openstack/openstack Project: openstack/cinder  b0b5f995546ef90a581eadb7558bee9c10b05b7b
VMAX driver - Does not honor clone volume size

It's possible to create a volume from a source volume
with a different size. The create_cloned_volume call
into the driver could have a larger volume['size']
than the source volume. So the driver will need to
create the clone, then check if it needs to extend the
volume to a new size.

Change-Id: Ie1c66cfb71c4761aa4b4b46e9a507c3acbfcbb1d
Closes-Bug: #11554773