Updated openstack/openstack Project: openstack/keystone  06ae44f3f1cd5be0d7acdf6bf7e33f3ae11610f1
Rename sql.migration_helpers to sql.upgrades

"Helpers" is one of those terms like "utils" that means "we don't know
what belongs in this module and what doesn't," so let's get rid of it in
favor of a shorter name that better reflects the purpose of everything
in the module. This name nicely opposes to sql.core, which contains
everything we use for more general database interactions.

Change-Id: I65ed77bc597b66eaffe5126275366f8ae141524b