Updated openstack/openstack Project: openstack/tempest  c0aa7ed49700114fe66f66f6f5d0f2378d0adb8b
Merge the separated link lines on service client

On current docstring of service clients, every link to API
reference is separated to 2 lines. That is unreadable on the
code and the documentation which is created from the docstring
also will be unreadablebecause of the separated link lines.

This patch merges the separated lines into a single line for the
readability.

Co-Authored-By: Shu Muto <shu-mutou@rf.jp.nec.com>
Co-Authored-By: Reedip <reedip.banerjee@nectechnologies.in>
Co-Authored-By: xu-haiwei <hai-xu@xr.jp.nec.com>
Change-Id: Ib1a4f2232dad8a2bc9d21a7c3509bc3b7bc0c827