Updated openstack/openstack Project: openstack/neutron  931c59f428c00b34a2d39c17d0c4ef004f232ed3
Skip warnings during DHCP port actions if no error is raised

With multiple concurrent DHCP operations, the server log may be
filled with warning traces about the inability to complete the
DHCP action correctly. These should be emitted only when reraising
the underlying exception, namely when the underlying error is
not an expected nor a handled one.

Related-bug: 1621345
Related-bug: 1620864

Change-Id: I1f00609f8181e3765238550f45e6e114d2df7e6e