Updated openstack/openstack Project: openstack/neutron  ba9b140e27bf19c4556060b7f9ea698e227ba109
Fix metering-agent iptables restore failure

When we removed the trailing space from iptables rules, we
broke the code that adds/restores them since it was assuming
it could split() based on that space to get the rule.

Just detect if the two values are equal, denoting there was
no rule instead.  This is fine since a self-referencing rule
is invalid anyways.

Change-Id: I9157ced10de3099790dea7f94aa4e614ebf7f25c
Closes-bug: #1623958