Updated openstack/openstack Project: openstack-infra/project-config  87fd919e1c4895ca1ffe43b3df14b8264fdf38ea
Include a marker file at the root of all docs builds

This is so that we can create a list of directories to ignore when
we rsync documentation builds to a publishing site.  Some builds
are published as subdirectories underneath the results of superior
builds.  This allows those superior builds to be rsynced without
removing the other builds that were published under them.

Change-Id: Icb0c02dc5b6f7b5e248e0df6d6093c29535b08f3