Updated openstack/openstack Project: openstack-infra/project-config  0f45635c0f6f3b28c727964da215bf6336f77ae3
Add root-marker for api-ref, api-guide,install-guide jobs

This is so that we can create a list of directories to ignore when we
rsync documentation builds to a publishing site. Some builds are
published as subdirectories underneath the results of superior builds.

This allows those superior builds to be rsynced without removing the
other builds that were published under them.

Also, create a common builder add-docs-root-marker and use it where
possible and update publishing of release notes as well.

Change-Id: I34cd631eb199d77801ad9fa068aa209f02ad1cfd