Updated openstack/openstack Project: openstack/openstack-doc-tools  54aa873505e7fab2ed76dadf1fa2e685fe1d66f1
Add root-marker file to translated manuals

This is needed for publishing in the future:
This is so that we can create a list of directories to ignore when we
rsync documentation builds to a publishing site. Some builds are
published as subdirectories underneath the results of superior
builds.

This allows those superior builds to be rsynced without removing the
other builds that were published under them.

http://specs.openstack.org/openstack-infra/infra-specs/specs/doc-publishing.html

Change-Id: Ic8773caebecafa87b10c1ad4612bde6b1513b667