Updated openstack/openstack Project: openstack/neutron  7b157b12e92eff5cc619f975157c9918c08d3013
Stop olso_messaging from error logging CallbackNotFound

This is expected when the agents are asking for a resource
that might not exist if a server side service plugin is
not loaded (e.g. Trunks). The Linux Bridge implementation
currently checks if each new port is a trunk so this results
in a bunch of log noise on the server without this patch.

Partially-Implements: blueprint vlan-aware-vms
Change-Id: I503fe51b0b537003a547af6e07304b29a53cde42