Updated openstack/openstack Project: openstack/neutron  827fd459fae9c814b74388194262425d6f26a0bf
Add metadata proxy router_update callback handler

This patch implements the callback handler for router update events;
This checks if the proxy process monitor is active, and if not, starts
the proxy.

This is particularly important if the metadata driver misses to receive
a create notification due to failures, which in turn generates an update
event because of a resync step.

Closes-bug: #1623732

Change-Id: I296a37daff1e5f018ae11eb8661c77ad346b8075