Updated openstack/openstack Project: openstack/tripleo-heat-templates  898902e411a3cc90d609603e9d264889bc007f63
Set client protocol for glance registry client

To communicate to glance registry, glance API has several parameters
that it uses to form the URI. Right now we are defaulting to http,
when we enable TLS everywhere, this will break. So setting the value
from the endpoint map should fix it.

Closes-Bug: #1623477
Change-Id: Id86787cbaa6f87fdcf9c26111c228fd59fbba012