Updated openstack/openstack Project: openstack/nova  a2ada0cabcbe42a81e38151d7ffba28af4e73d0d
HyperV: ensure config drives are copied as well during resizes

During cold migration, vhd config drive images are not copied
over, on the wrong assumption that the instance is already
configured and does not need the config drive.

For this reason, migrating instances using vhd config drives
will fail, as there is a check ensuring that the config drive
is present, if required.

This change addresses the issue, removing the check that was
preventing the config drive from being copied.

Change-Id: I8cd42bed4515f4f75c92e595c4d8b847b16058dd
Closes-Bug: #1619602