Updated openstack/openstack Project: openstack/ironic  25fcc07507331c3ed7928dbe31f8a231797b00f3
Mark untested drivers as unsupported

This will cause them to log a warning at startup, that it is unsupported
and may be removed later.

The drivers marked unsupported are any driver that uses:

* SSHPower / SSHManagement
* NativeIPMIPower / NativeIPMIManagement
* seamicro.Power / seamicro.Management
* IBootPower
* SNMPPower
* VirtualBoxPower / VirtualBoxManagement
* AMTPower / AMTManagement
* MSFTOCSPower / MSFTOCSManagement
* WakeOnLanPower

See the release note in this change for the full list of the setup.cfg
names for the drivers.

The remaining drivers are tested via OpenStack CI, and/or have a
third-party CI listed here[0] and are making an honest effort toward
having stable CI.

[0] https://wiki.openstack.org/wiki/ThirdPartySystems

Closes-Bug: #1526410
Change-Id: I9c09be6b7a734426a4c0b18cb61ece5389428cd5