Updated openstack/openstack Project: openstack/openstack-manuals  531118e545258bf880acd315db2f923416cea93c
[install] Update keystone for Newton

This commit removes the ADMIN token steps from the install guide and
replaces them with ``keystone-manage bootstrap``, which is the
preferred method for bootstrapping a keystone deployment without
the security risks of a privileged administrator token.

This commit also introduces an additional ``keystone-manage`` command
to setup a fernet repository for encrypting and decrypting
credentials.

Change-Id: Id1f490e61aa357f3fc9b71307196835410c0d08a