Updated openstack/openstack Project: openstack/neutron  8a03808ebc61880726df948cadd6a6d3a57e5a4a
Demote error trace to debug level for auto allocation operations

Based on how Nova calls into this extension, it is clear that this
'error' shows during normal conditions of booting  multiple VMs for
the same tenant. Plus, there is nothing the operator can do as the
plugin hopefully takes care of it all.

Change-Id: I6747e64903f8e57ddadcf2d1c74d223fb4fdf97d