Updated openstack/openstack Project: openstack/swift  6fea41e0f63739428559dd1e3396f2b04b369bac
Updating Bandit config file

This commit updates the Swift Bandit file to the new style
introduced in Bandit 1.0.  In response to the struggle with
getting a Bandit config file working and kept up to date we
introduced a simplified version in Bandit 1.0.

This commit updates Swift's bandit.yaml to use the new version.

Change-Id: Ida5dd08f4ea72a377346f2159caeb2f3741d4980