Updated openstack/openstack Project: openstack/ironic  6782f9c98ba4cd1e5d1607e37d45d6ba429c8bd5
Remove NotificationEventTypeError as not needed

The exception NotificationEventTypeError was used when an invalid
status was used. But it is not possible to assign an invalid status
without causing a ValueError.

Add a unit test to show that changing status to an invalid value does
raise a ValueError.

Remove check for valid status value.

Change-Id: I58fb50f963ca4a8e8d0185b9165c9bcda3bb7a25