Updated openstack/openstack Project: openstack-infra/python-storyboardclient  d7622a8f0a44af0076f5a1c7e252d622ba9d6f35
Update 'usage'

This updates the commentary on the python client, now that it can
work with StoryBoard instances that use self-signed certificates.

It also removes 'examples of delete commands' from the TODOs, as
we don't want to make it easy for passers-by to copy-paste a command to
delete data in bulk without any understanding.

Change-Id: Icb9660ac79992a1e53baa89f1016c626b0cfe467