Updated openstack/openstack Project: openstack/neutron  e1a29a6ff4c65cf64d89fd79970b5ce05c7e06c9
Don't warn in VLAN Plumber on tagless children

The VLAN Plumber VLAN parsing was warning on lines that had
parent interfaces but were missing 802.1Q tags. This was warning
on veth pairs and bridge interfaces, which was silly. This
adjusts the logic to handle those without warning.

Partially-Implements: blueprint vlan-aware-vms
Change-Id: I4a78e7bce2883f6ca77c18bd388c03c99b8d0ae1