Updated openstack/openstack Project: openstack/heat  3af406f482bc2248794416bc7cfca16355aaf5c1
Fix sync point delete

When a resource failed, the stack state was set to FAILED and current
traversal was set to emoty string. The actual traversal was lost and
there was no way to delete the sync points belonging to the actual
traversal.

This change keeps the current traversal when you do a state set, so that
later you can delete the sync points belonging to it. Also, the current
traversal is set to empty when the stack has failed and there is no need
to rollback.

Closes-Bug: #1618155

Change-Id: Iec3922af92b70b0628fb94b7b2d597247e6d42c4