Updated openstack/openstack Project: openstack/nova  fe46d482d6342c82a76971a31012b0a00be5fb66
Switching expression order within if condition

In nova/cmd/manage.py, on line 294 [1] there is the following line:

  if value['limit'] < 0 or value['limit'] is None:

The problem occurs when value['limit'] is None, as this line tries
to compare None (object) with an integer value.

This comparsion is totally okay (without errors) in Python27.
But, if you run this on Python34, a TypError is raised:

  TypeError: unorderable types: NoneType() < int()

It is simply fixed with swapping the two expressions within the
if condition. Now if value['limit'] is None, the first expression
becomes true and second one does not get evaluated.

TrivialFix

[1]: https://github.com/openstack/nova/blob/master/nova/cmd/manage.py#L294

Change-Id: Id2b1ed67e1c2c9f1d1aa5a23740db2b18e2e7141