Updated openstack/openstack Project: openstack/nova  96ab6c91cd163b9803c279a70d3db28c412afbce
Fix database poison warnings, part 12

The following warning appears in the unit test logs a number of times.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

This patch fixes the last two warnings from:

    unit.compute.test_compute_api.py

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: I8fefd8228f2aa9490cab943a79c50e76b1e66fa8
Related-Bug: #1568414