Updated openstack/openstack Project: openstack/cinder  205a850276b3763462dcb77f8ed4f185da97e8d0
Dell SC: ISCSI initialize_connection fixes

The driver could return duplicate iSCSI portal/IQN/Lun combinations.
This patch returns the condensed list.

Also sometimes when Cinder attempts to attach to a volume it will use
the first item in the multipath list rather than the specified single
path entries. So this patch simply moves a known up and active entry
to the front of the list.

Change-Id: Idb371825b6520d4bdb09a1dbb37429cb10643e93
Closes-Bug: #1622703